+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - Change `SearchInput` to have a label, for accessibility purposes
+- Use new credit card promote endpoint
 
 ## [2.33.0] - 2024-12-02